star_tides/requirements.txt
flask
psycopg2-binary
sqlalchemy
alembic
flask_sqlalchemy
flask-mongoengine
mongodb-migrations
celery[redis]
bcrypt